Abortive attempt at refactoring JavaBeanUpdatableCollection to always use IUpdatableValue to update the collection.  Here are the problems:
1) IUpdatableValue needs a property of an object, which doesn't work with collection elements, so we need to create a collection of objects with properties that IUpdatableValue can edit.

2) As soon as we do that, we cause two more problems:

a) We have to make sure that whenevery anyone in the application edits that same collection, it always listens to exactly the same set of IUpdatableValues, or updates won't work across collections.

b) We hard-code a non SWT.VIRTUAL implementation.

Deferring implementation until we can talk about the API implications of this.