Since jsch 0.1.31, it has its own order to try authentication methods and it is simmilar to OpenSSH's "PreferredAuthentications" option.
By this change, keyboard-interactive(KI) auth method will be tried and then
password auth will be tried.  Such a behavior is not a problem and, in fact
OpenSSH's ssh will have same behavior, but unfortunately the prompt for
KI[1] is not so matuare and it will not allow to save given password.

By this commit, password auth will be tried if possible and then KI will be.

[1] org.eclipse.team.internal.ccvs.ui.KeyboardInteractiveDialog.java