Bug 352076 [Compatibility] Exception caused by having the 'XPath' view open prevents Ctrl+W from operating
Part listeners were not being notified safely so if an exception
occurred in one of the listeners it would cause all other
notifications to not occur. The fix then is to wrap notification
calls within an ISafeRunnable so that every listener will get a
chance to be notified even if another listener causes an exception.