Bug 352462 [Compatibility] F1 does nothing in the 'Git Staging' view
The 'Git Staging' view as-is does not provide any contextual help.
In 3.x, hitting the F1 key will bring up the help for the workbench
window instead. To get the same behaviour in 4.x we need to hook up
the contextual help to the workbench window's shell so that hitting
F1 in the view will get the 'Help' view to show the same content it
is showing in 3.x.