Bug 352508 [Compatibility] Closing active perspective should set the window to the last active perspective
The code was originally just closing the perspective and if it was
the active perspective, the model service would set the selected
element of the perspective stack to the first available perspective
in its list. Instead of this arbitrary selection, the window should
instead be selecting the previously active perspective. The fix is
to update the perspective stack's selected element prior to letting
the model service close the perspective.