Bug 353229 [Compatibility] serviceFactory extensions are being ignored
The current service locator only queries the context for
services which means that any service factories contributed via
extension points will not be processed. The fix is to first
query the context and then go through the local service factory,
the extension registry, and parent service locator for services
that could not be found in the context.