Bug 353248 [Compatibility] Disposable services created by the context may not get disposed
Services that are created by the context do not get disposed
even if they are implementing the IDisposable interface. If the
context returns a service that is created locally and this is
retrieved through the service locator, the service locator
should record this so that the service can be disposed by the
service locator when the service locator gets disposed.