Bug 353248 [Compatibility] Disposable services created by the context may not get disposed
The logic for determining whether a retrieved service had to be
disposed when the service locator got disposed was incorrect.
This was because the get(String) method was being invoked on the
context instead of the getLocal(String) method. This meant that
the context would recurse upwards to search for a service when
it should have only been looking at its own local cache.