Bug 352905 [Compatibility] 'Quick Outline' disappears if opened after opening an editor
As the stack renderer uses asynchronous activation requests to
activate stacks, the requests can potentially be invalidated
when this enqueued request actually gets processed. In the case
of this bug, a dialog window gets opened in the interim time
between the activation request being enqueued and when it
actually gets processed. This causes the dialog to lose focus as
the request gets processed and causes the stack itself to be
activated. The fix is to not activate the stack if it has
determined that it is not currently a part of the activation
chain.