Bug 354739 [Compatibility] Perspectives with standalone views lay out incorrectly
When parts need to be placed relative to another part, the part's
parent should be considered if it is a stack as the parts cannot be
placed next to a part inside a part stack. However, if a part's
parent is not a part stack then the part should just be placed
relative to the part.