Bug 355210 EHandlerService's canExecute(*) should not propagate exceptions upwards if an error occurs during enablement evaluation
At the moment, the code directly asks the handler whether it is
enabled or not without any regard of any exceptions that the
underlying implementation throws. This causes problem for
contribution updates as one rogue implementation will cause all
other succeeding contriubtions to not get updated as the
propagated exception will intervene with the current code path's
execution. The fix is to protect the enablement check and return
'false'. The exception will only be logged if tracing is enabled
to prevent the log from filling up.