Bug 355056 [Compatibility] NPE in MenuHelper$3.compute on startup with Juno Java EPP package
Editor actions may be inserted into the model on startup but
they should not be visible on startup until the relevant target
editor is up. Contribution items should have its visibility
state determined before being inserted into the model to prevent
being initialized when the renderer when they shouldn't be.

It is not possible for editor tool bar contributions to get
information about its parent part as the contributions are
inserted into the workbench window's tool bar and the editor is
not available in the window's context. The fix is to just
initialize the action descriptor without a target like it is
done in 3.x.