Bug 354967 [Compatibility] Menu items appear where they shouldn't with Eclipse 4.2
Contributions that only had a 'visibleWhen' clause using a
'checkEnabled' for commands were not working. This is because
the code for the processing only generated an expression if the
clause actually had children. In the case where there is only a
'checkEnabled' attribute pointing to 'true', the element would
have no children. The fix is to consider this case and generate
an expression which evaluated whether the handler was enabled or
not.