Bug 356124 [Compatibility] View is not displayed if added relative to a placeholder view
When a stack that should be rendered is placed under a part sash
container which does not need to be rendered, the part sash
container should change its setting so it does get rendered.
Likewise, all of its parents should also be rendered to ensure
that the newly added stack can be seen in the perspective.