Bug 355964 [Compatibility] Tool bars are duplicated in the main workbench window
The main tool bar is added to the window's trim the first time
an update is requested but the check there only checked for
whether the trim bar already contained the main tool bar. During
the shutdown phase we remove all trim elements from the trim bar
so the trim bar is once again empty. In the event that a user
requests an update to the window's action bars the main tool bar
will be added to the trim again, causing there to be two tool
bars on the next workbench startup. The fix is to only ever
perform the add action once.