Bug 358566 [Compatibility] ViewDescriptor does not implement isRestorable() correctly
The code originally assumed that view contributions that did not
explicitly define its 'restorable' attribute should not be
restorable. This is not correct as the default should be 'true'.
The fix is to retrieve the string from the extension registry
first and only parse it for its boolean value if the string is
actually non-null. If it is null then 'true' should be returned.