Bug 358204 [Compatibility] Huge toolbar buttons with EcoreDiagramEditor
The ImageDescriptor parsing code that was originally implemented
only checked for URLImageDescriptors. GEF editors call
ImageDescriptor's static createFromFile(Class, String) method so
code also needs to be added to handle FileImageDescriptors.
Reflection is used to access the private fields contained within
FileImageDescriptor and from that information a URI is
constructed with a bit of string manipulation.