Bug 351358 [Compatibility] Toolbar ordering is off in 'Git Repositories' view
No model tool items were being recorded for individual
contribution items that were being added to a view's tool bar
manager. As there was no model representation of the items, the
order in which the tool items were being laid out were incorrect
in the cases where there were both programmatic and declarative
contributions to a given tool bar. The fix is to create opaque
tool items for the programmatic contributions so that they will
be represented in the model.