Bug 359900 [Compatibility] Unhandled URL protocol "platform" on GIF files
The view registry was automatically prepending the special
"platform:/plugin/" to all 'icon' attribute values that were
read from the registry. This is not correct as some icons may
already use this special URI as their icon definition. The fix
is to check if the definition is already there and then skip the
processing step if it's already there.