Bug 362582 Focus is lost after switching perspectives on OS X 10.6.8
When a perspective becomes activated, its controls may not be
visible yet because the layout has not yet happened. In this
case, the @Focus invocation essentially becomes a no-op as the
windowing system may decide not to grant a control that is
invisible focus. The fix is to have the perspective stack
relayout its children first before activating its context chain.