Bug 340771 [Compatibility] Debug views are empty when first opened
As the part visibility events are only fired when the selection
of an element container changes, the event may not be fired on
startup because there have been no changes. This causes the
debug views to show up with no content because they optimize
themselves to not render anything if they are not currently
visible. The fix is to fire these visibility events on workbench
startup.