Bug 364757 [Compatibility] Action set contexts can be deactivated too eagerly
When a perspective switches, we should go through the action set
manager to show and hide action sets. Going directly through the
EContextService means that the code is out of sync with other
code that is using the action set manager. Having the manager
show and hide action sets ensures that repeated calls to show or
hide will be cached appropriately so allow the correct
visibility state to be calculated.