Bug 348920 [Compatibility] Contribution visibility seems to leak to other workbench windows
Currently, when a workbench window is asked for an
IContextService, it returns the one that the application owns.
This means that any context activations and deactivations will
be done on a global scale. This ultimately means that changes in
one window's context may affect the context in another window's.
Every window already has its own EContextService implementation
so if it also has its own IContextService implementation,
changes to the context will only affect that one window instead
of the entire application.