Bug 349104 SWTException thrown after changing perspective and closing the window
When a perspective is changed, trim stacks may appear or
disappear which means that the shell must have its layout
recalculated to accommodate the new controls that have been
created. In the event that the window is closed before the shell
gets a chance to layout, the layout operation should just be
skipped as the shell has since been disposed.