Bug 319621 [Compatibility] All views now have view menus
The rendering of a part's menu is a very dynamic problem given
that the decision to have a menu may vary based on a wide range
of variables. A part have a menu when it is first constructed
but it may no longer be needed later. Conversely, a part may be
instantiated without a part menu but one may need to be created
in response to some change.

This change adds code into the updateActionBars() call to ask
the renderer to recalculate whether a view menu should be
rendered or not. This allows the menu to be generated or
disposed dynamically based on the current state of the
application.