Bug 364073 [Compatibility] "Compare With" always done on same file regardless of selection
IHandlerService's getCurrentState() needs to return an
evaluation context implementation that wraps our actual
evaluation context that's backed by the eclipse context.
Otherwise, if we return the original backing version, then any
changes made by the client to the returned evaluation context
will modify the underlying eclipse context. The user should not
be allowed to cause the system to change like this.