Bug 365202 NPE in ContributionRecord's updateVisibility
When a part is disposed, we remove references of its menus from
the menu manager renderer to prevent leaks. However, the context
may be modified in some way after this is done so the
contribution record may not be able to retrieve its target model
element's menu manager. Just check for this case if it is
encountered and ignore it.