Bug 367963 [Compatibility] 'Debug' view is unusable with CDT installed
When a part's toolbar gets disposed, its manager's link and the
renderer is removed. Afterwards, the part's context itself is
disposed. At this point, a RunAndTrack of the toolbar is
notified of this change. Since the toolbar itself has had its
links cleared, the toolbar should ignore this processing and
also remove its RunAndTrack from the system.