Bug 368751 [Compatibility] NPE on startup due to null editor site
If an editor's input cannot be restored, an error editor part
should be displayed to the user to indicate this. However, the
current code chokes again because it attempts to load the
original editor's input and set it as the error editor part's
input. Instead, a NullEditorInput should be created for the
error editor part when the original editor input cannot be
restored.