Bug 349423 [Compatibility] 'Source' and 'Refactor' menus disappear after installing Photran
When two menus with a matching element id are found, the cloned
copy should not simply be discarded without regard to any of its
visibleWhen clauses. Instead, its visibleWhen clause should be
merged with the one that was found so that the menu will be
shown to the user if either of the two conditions have been
satisfied. Otherwise, the menu will only ever be shown if the
first condition is satisfied regardless of whether the second
condition has been satisfied or not.