Bug 369790 [Compatibility] Views with secondary ids do not get opened correctly in another perspective
Our search rules for views currently only consider the active
perspective. This is obviously incorrect because a view can
exist in multiple perspectives. The only reason this ever worked
before was because of some code that allowed views without a
secondary id to not be duplicated.

The fix is to have the search go through the entire workbench
window and all of its perspectives so that parts that exists in
hidden perspectives are found and instead of being duplicated.