Bug 370026 NPE in PartActivationHistory.activate() in almost fresh 4.2M5
When the model is being rendered and a piece of code tries to
"interrupt" the process by forcibly showing a part in another
part of the model that has not yet been rendered, the request
will fail because the rendering engine may not be able to create
the element if its parent hierarchy do not have their widgets
set. The fix is to forcibly create any unrendered parent in the
hierarchy chain to ensure that the explicit request to create
the part itself will succeed.