Bug 370441 Allow to see all editors in ">>" drop-down
Update the code so that clicking on a CTabFolder's chevron will
bring up all the editors in an order dictated by the MRU
setting. If MRU is on, the list will be shown alphabetically. If
MRU is off, then the tabs that are in the folder will just be
shown as-is with no reordering.