Bug 348069 Closing an additional window of an Eclipse 4 application doesn't unrender it
When a window was being closed, only its widgets were destroyed
and its context structure was not. This leads to memory leaks.

If a top-level window gets closed, its entire structure and
context will be destroyed and the window itself will be removed
from the application's list of child windows if it is not the
last window in the list. If it is the last window, then the
window's structure and contexts will be destroyed but it will
not be removed from the application's list of child windows for
restart purposes.

For detached windows, detached windows will have always have
their structures and contexts destroyed but it will only remove
itself from its parent list of child windows if the detached
window being destroyed does not contain any parts.