Bug 319621 [Compatibility] All views now have view menus
A part's menu is a dynamic item in that when the part is first
constructed, it may not have a menu but a menu may become
necessary at a later point in time when other variables become
satisfied. Likewise, a part may start of needing a menu but this
requirement may be lifted later.

To properly react to this, a RunAndTrack is needed to determine
when a part's menu should be rendered or not.