Fix Bug 363334 - SWTBot test launcher doesn't work with Eclipse 4.2
We copied TestableObject and ITestHarness into
org.eclipse.e4.ui.workbench3, but kept the copies in
org.eclipse.ui.workbench.  This could lead to LinkageErrors depending
on the load order of e4.ui.workbench3 and ui.workbench.

A reproducible test case seems to be the MyFirstTest from their
EclipseCon 2009 example with SWTBot 2.5.0
<http://wiki.eclipse.org/SWTBot/UsersGuide#Getting_started_with_SWTBot_for_Eclipse_Plugins>