Bug 371611 LinkageError in DynamicToolBarContributionItem can cause workbench startup to fail
If the createControl(Composite) method of a ControlContribution
returns null, we should not bother trying to create a ToolItem
for it. Instead, we should just log an error notifying the user
which class failed to return a non-null control. The current
code of throwing an NPE causes the workbench to not start when
there is such a rogue control contribution. This is unacceptable
as a bad implementation should not prevent the workbench from
starting up.