Bug 371817 [Compatibility] 'Close All' stops working
Switch the enablement to be determined by the active part and
active perspective. This will mean that the handler will still
be enabled even if all the editors have closed if the active
part doesn't change but this is better than it being disabled
and being next to impossible to enable it back due to its
current dependency on the active workbench window.