Bug 355059 [Compatibility] Mylyn's activate task trim is not shown
As the renderer doesn't correctly handle models that contains
multiple trim bars with the exact same side values, trim
contributions to org.eclipse.ui.trim.command2 needs to be placed at
the end of org.eclipse.ui.main.toolbar.

To achieve this, the "endof" locationURI needs to be supported for
starters. The contribution processing must also not happen before the
perspective switcher and other elements are appended to the trim bar
or else the trim contributions will be added before these other
elements when they should be at the end.