Bug 366568 [Compatibility] command service refreshElements(*) not implemented
In order for handlers to be notified of a refresh, the
MakeHandlersGo needs to delegate update requests down to the
real handler that it is wrapping. Since CommandContributionItems
are being converted to HandledContributionItems,
HandledContributionItems need a way to hook themselves up for
updates with the 3.x ICommandService. This is done through the
IUpdateService service interface that acts as a bridge between
the two systems. An implementation is provided from the
org.eclipse.ui.workbench that will consume and convert
registration requests into 3.x registration requests so that the
two systems are able to talk to each other without the 4.x code
having an actual hard dependency on the 3.x code.