Bug 367094 [Compatibility] Programmatically created CommandContributionItems show wrong enablement state
When the handler that the proxy is wrapping has fired an event,
the proxy should notify MakeHandlersGo about the event so that
the command itself will be notified of the event. Since
CommandContributionItems listen to the command for change
events, this wiring will get the CCIs to react properly to
handler changes.