Bug 319704 - Toolbar button ordering broken
Allow the actionSets to float on top of the Eclipse4 menu and toolbar
renderer.  The contributions visibility and placement is then managed by
the 3.x legacy actionSet code.


Don't show any action sets during the initialization process.
Doing so causes multiple show requests and subsequent hide
requests will not be sufficient to get an action set to go
hidden due to reference counting.

The temporary perspective did not have an id set on it so
getPerspective(MPerspective) code was unable to identify which
perspective descriptor it represented. This meant that it had no
action sets to activate which causes an action set to be
deactivated twice. This makes the reference counts go into the
negatives and causes them to pretty much never become active
again.

Bug: 319704