Bug 378845 - Customize Perspective not presentable on 4.2
With the changes in 4.2 workbench, the 3.8 CustomizePerspectiveDialog is
much closer to the desired behaviour than the 4.x version.  This starts
by working with actionSets.  It still needs to be modified so the first
2 tabs work with the model.

Implementation that uses the window Eclipse4 model to display and hide
the main menu.

CustomizePerspectiveDialog modified to manage toolbar visibility from
the window top trim.