Bug 388966 [Min/Max] [Trim] CCE on Maximize/Minimize
Depending on how the animations preference is set, the value that is
returned by the event may be of a different type. A Boolean is provided
if the change is from the preference page directly and a String is
provided if it was modified because of the preference has been
imported. The fix here is to check for each of these two cases and set
the value the event represents as a Boolean.