bug 391086: Search the model editor tree. https://bugs.eclipse.org/bugs/show_bug.cgi?id=391086 The FilteredTree, PatternFilter and BasicUIJob are already copied in org.eclipse.e4.ui.workbench.swt *.internal.copy package so instead I now depend on this bundle.
I completely removed all the code from the existing PatternFilter
(sorry) and let it implement the "original copied" PatternFilter from
org.eclipse.e4.ui.workbench.swt and instead only keep the method that
handles with the ILabelProvider. 

So, I am checking in what is now a minimal change so that you can give
it a spin. 

@TODO: Take care that the tree does not collapse when search is
cancelled 
@TODO: Wrap FilteredTree in ShadowComposite so that it looks like it did
(takes a lot of space IMHO)