[Edit] Enable center arrow buttons when editable and for both sides
With this, the [>] and [<] buttons can also be used for non-conflicting
situations, just like the toolbar buttons "Copy Current Change ..." can.
It's useful e.g. in EGit to partially stage changes.

It also enables using them when both sides are editable, by showing a
[>] button and a [<] button below it. In the compare editor opened from
EGit's Staging view, this is the case where the left side is the
workspace file and the right side is the index.

Bug: 398594
Change-Id: I392b181fc4bc7e31502e6edf9b59e4b8b27b8074