Bug 417726 - [Themes] Default theme toolbar gradient looks bad on GTK
This changes the toolbar from an uneven image of a gradient to a smooth
gradient from the default widget color to the background color used in
the workbench. It makes it much more pleasant to look at and is better
integrated with the look of other applications. See bug for screenshots.

It also changes the perspective switcher gradient to match; it started
with a slightly different color before.

Note that this depends on the change for bug 420035 in platform UI.

Change-Id: I7d2c1297efa1e5b308853d7dd2afc7f06c1f4264
Signed-off-by: Robin Stocker <robin@nibor.org>