bug 422289: [IDE] copying/pasting elements from the Markers view should preserve their listed order
Copies selected items in the order they appear in the view. HOWEVER,
when a category is selected, and the view displays only a portion of the
markers in the category because the views display limit, only the
displayed markers are sorted. The remaining markers are included in the
copy after the sorted ones, but in no definite order.

Change-Id: I1a205b50a789039a6dfcb29dabeab95717e538f4