Bug 436985 - [Backport] activePart within SelectionAggregator is not set to null if all parts are closed
If all parts within the workbench are closed the method setPart of the
SelectionAggregator is called with parameter part = null. In this case
the active part has to be set to null. Otherwise there is an
inconsistent state of the selection aggregator if one part is reopened.
The selection of the newly opened part would not be tracked correctly.

Change-Id: I4adc486f5ba1e78bffbcc4ae3a8a333c596e3733
Signed-off-by: Oliver Puetter <oliver.puetter@dew-stahl.com>