Bug 423882 - Warn user if workspace is newer than IDE
Refine the workspace version check and show up
* a warning if workspace > IDE
* an info if workspace < IDE.

Compare against only major and minor version of the IDE, so that
for a change of just the patch segment no info is shown. As a reference
bundle use org.eclipse.platform since its feature is most likely to
change when the "IDE" changes.  Don't want to check against the feature
itself as that might not be installed at all.

Compatibility notes:
- If coming from legacy workspaces, we don't complain since that might
  be too intrusive.
- We also write an incremented legacy version, so that when going
  back from an updated workspace to a legacy workspace the old IDE would
  complain as well.
- If org.eclipse.platform bundle is not installed, no check is
  performed.

Change-Id: I2ef3053d4a61782315c95543c5d7b535e0bef0a6
Signed-off-by: Christian Georgi <christian.georgi@sap.com>