Bug 426785 - [Help][Search] Querying help content while exiting Eclipse may cause a deadlock preventing a proper shutdown
In org.eclipse.help.internal.search.SearchIndex deadlock removed: The
method "close()" waits for all searches to finish. But the code to wait
was enclosed by the same "synchronized (searches) {...}" as in
"unregisterSearch(Thread t)", which must be called to finish a search.
To avoid the deadlock the code to wait has been moved outside the
"synchronized" block.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=426785
Signed-off-by: Holger Voormann <eclipse@voormann.de>
Change-Id: Ib3ffbe4e7c75ad7b6bcb359cb8788650f2eabbf7