Bug 410426 - VisibleWhen from contribution factories not handled in toolbars
ToolbarContributionRecord did not handle the factory case in
anyVisibleWhen() and always return false for contributions through
org.eclipse.ui.menus.ExtensionContributionFactory.

This commit also correct a ClassCastException in
ToolBarManagerRenderer.toBeRenderedUpdater EventHandler: the
IContributionItem is not always a ContributionItem, it could be a
MenuManager too. 

Bug410426Test contains two tests which checks the corrected behaviors.

Change-Id: I5fd8b73ab8224fe9bd549549fac5eb0daf1b322f
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>