Bug 364324 - [Help][Search] Model result categories ordering after TOC ordering
Problem:
By default the search result categories are on a par with the top-level
books (TOCs), but the order of the TOCs and the order of the result
categories can differ. For example, if you search for "subsystem"
(http://eclip.se/1r) and "Show result categories" is enabled, then the
category "Workbench User Guide" is shown in the search result list at
the bottom, in contrast to the table of contents where the "Workbench
User Guide" book is shown at the top. From the user perspective this is
confusing.

Caused by:
Both the books in the Contents tab and the categories in the Search
Results tab are by default alphabetically ordered. The table of content
book order can be changed via the Help Data file (see
http://eclip.se/1s), for instance in the Eclipse IDE the "Workbench User
Guide" is shown at the top. In contrast, the search results category
order cannot be changed.

Solution:
The search results category order should follow the table of content
book order.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=364324
Change-Id: I04a3eb4df1b75db5459dbafb4e4c9745c00d548a
Signed-off-by: Holger Voormann <eclipse@voormann.de>