Bug 430662 - Editor area splits when dragging an editor without Ctrl key pressed
Don't split the editor area when dragging within the editor area and the
Ctrl key is not pressed. Don't split the editor area when dragging into
the editor area from outside and the Ctrl key is pressed. Show outer
rectangle as an indication that the part will be inserted into an
existing editor area stack. 

Change-Id: I3df178aa8ff56c47ee71bbd0489cd4be05deaa04
Signed-off-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>