Bug 437005 - Out-of-date .snap file prevents Eclipse from running
The fix changes the name of the tree snapshot file from .snap to
<sequence_number>.snap, where the sequence number matches the sequence
number of the tree file. With this change a failure to delete a snapshot
file after creating a new tree file is no longer fatal since the
leftover snapshot file would not match the sequence number of the tree
file and will be ignored while restoring the workspace.

Backward compatibility is achieved by checking for the .snap file if the
<sequence_number>.snap file is not found. The .snap file is accepted
only if its last modification time is not lower than the last
modification time of the tree file.

Change-Id: I2a8ca8ad79b788089e4d42d1c310f642f152700a
Signed-off-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>