Bug 404348 - remember disabled action sets
The action set "visibility" state managed by ActionSetManager was never
persisted and the user choice to disable action set was lost on
shutdown. This change persists user choice in the same way as it was
done for toolbars and menus.

Note: plugins can request to show/hide ActionSet (depending on the
application needs), but once user decides to "disable" action set, his
decision must override application choice. The best example is the debug
action set with the global debug buttons which requests "visibility" on
each breakpoint hit. The change allows Eclipse to ignore those requests
once user decided to do so.

Change-Id: Ie3a962b42f2e157344385e4a16bcc8782d64c1c6
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>