Bug 420956 - Fix perspective customization on 4.x
The CustomizePerspectiveDialog shows now all toolbar/menubar
contributions from action sets (with right labels) previously shown in
3.x stream, allows disable/enable button status per perspective
and persist the state in the model. Additionally the label provider is
enhanced to show the availability of almost all contribution depending
on the action set state.

Note: disabling and re-enabling entire toolbars is still broken as
described in bug 383569 comment 24 and will be fixed in the next
patches.

Change-Id: I6f002c518e671437775187ffcc6c51050aa8e0ae
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>
Also-by: Erik Chou <ekchou@ymail.com>
Also-by: Paul Webster <pwebster@ca.ibm.com>
Also-by: René Brandstetter <Rene.Brandstetter@gmx.net>