Bug 457687 - Dependency Injection should not catch Errors
According to the JavaDoc, an Error is "a subclass of Throwable that
indicates serious problems that a reasonable application should not try
to catch". Re-throw Errors that are raised by methods invoked by DI.

The underlying issue is that the Reflection API wraps Errors in an ITE.
It has been admitted in a JDK bug [1] that this is a bad practice but
cannot be changed for compatibility reasons. The recommended workaround
is to re-throw Errors wrapped in ITEs.

[1] http://bugs.java.com/bugdatabase/view_bug.do?bug_id=4386935

Change-Id: I109b64cc98b63e9852043fd232eb0664f3156ea4
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>