Bug 458261 - Error messages from early startup should be more descriptive
Additionally to the extra information from the extension the error
reporter id is changed to be from the platform ui, to make clear that
the error is reported by the platform and not by the broken extension
code. As a side effect this avoids the possible NPE during logging if
the wrong configuration element has no namespace (current code of
TableReader.basicLoadExtension() allows this).

Change-Id: If3218e07416ed2824bc2473e0b1c2f3a82f0e07c
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>