Bug 445538 - [Perspectives] Pin Editor command broken and can no longer be added to perspective's toolbar
Because CommandContributionItem's are now replaced with
HandledContributionItem's in CoolBarToTrimManager we can't update the
state of originally contributed items anymore.

Moved the "pin editor" show/hide and refresh logic from the anonymous
subclass of the ContributionItemFactory.CommandContributionItem into the
WorkbenchActionBuilder.

Updated CPD tooltips and popup text which warn user that selected item
visibility can't be changed via CPD if the item represents an initially
disabled CommandContributionItem (default is enabled).

Change-Id: I313290938b012c9578a05a999a05f76bac5e2bfb
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>