Bug 458704 - Fix failing FileSearchTest
Update the TextSearchRequestor's API to support parallelization
and fix existing subclasses and tests.

Also add an optimization to process all files in a single job if
the maximum number of threads for the job group is 1, which can
occur on single processor machines or if the TextSearchRequestor
passed to the TextSearchEngine.search() does not support
parallelism. This removes some unnecessary job scheduling.

Change-Id: I5baaf1dc3910a527338dc12e7fa4575269470833
Signed-off-by: Terry Parker <tparker@google.com>