Bug 410292 - IWindowCloseHandler and ISaveHandler are not easily exchangable
Installs only one instance of the IWindowCloseHandler and ISaveHandler
in the application-context instead of creating a separate instance for
every window. This makes it easier to overload them on a per window base
or replace the only available instance at the application start via the
UILifeCycle.APP_STARTUP_COMPLETE UI event.

Change-Id: I507ad731b4c7e867afb40cd7d407078f30e5a85f
Signed-off-by: René Brandstetter <Rene.Brandstetter@gmx.net>