Bug 470233 - Replace new Boolean with Boolean.valueOf
Using `new Boolean()` results in the creation of a new object on the
heap, when the flyweight `Boolean.TRUE` and `Boolean.FALSE` are available.
Java 1.4 added a `Boolean.valueOf()` which can be used in place of
`new Boolean()` but which will use the existing flyweight values instead.

Globally change `new Boolean(...)` to `Boolean.valueOf(...)` and replace
constant valued expressions with their flyweight counterparts. In-line
`Boolean.valueOf().booleanValue()` with `Boolean.parseBoolean()`.

Bug: 470233
Change-Id: I06be411c8e2ef387275af06409996acdab7c2cf3
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>