Bug 470244 - Replace new Boolean with Boolean.valueOf
Using `new Boolean()` results in the creation of a new object on the
heap, when the flyweight `Boolean.TRUE` and `Boolean.FALSE` are available.
Java 1.4 added a `Boolean.valueOf()` which can be used in place of
`new Boolean()` but which will use the existing flyweight values instead.

Globally change `new Boolean(...)` to `Boolean.valueOf(...)` and replace
constant valued expressions with their flyweight counterparts.

Bug: 470244
Change-Id: I608c1acf8b35274e3e8325de8d396f2d077001ce
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>