Bug 475417 - Optimise contenttype loading time
The existing Activator for the contenttype bundle was just registering an
instance of a service and connecting it to other services (such as DebugOptions
and IExtensionRegistry).

By moving those into a declarative services component, the framework will start
this service up asynchronously and the activation of the bundle no longer has
to run any code.

Bug: 475417
Change-Id: I5c4f060fae84cc9e0aaa60f5d613ba3102d450ab
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>