Bug 476364 - Remove BundleActivator from DI extensions
The Activator in the org.eclipse.e4.core.di.extensions bundle does very
limited things; it builds a set of listeners (that can be unsubscribed at
bundle stop time) and connects the user to the EventAdmin and
IPreferencesStore services. Given that the objects are instantiated by
Declarative Services already, use that to wire the services up instead.

Change-Id: Idd567f422caaabdf611182985bcbf3a65247e678
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>