Bug 476364 - Remove BundleActivator from DI extensions
The Activator in the org.eclipse.e4.core.di.extensions bundle does very
limited things; it builds a set of listeners (that can be unsubscribed at
bundle stop time) and connects the user to the EventAdmin and
IPreferencesStore services. Given that the objects are instantiated by
Declarative Services already, use that to wire the services up instead.

Change-Id: Id12eb6d9600ab43f66e27c2069d58a5f0a2ecb74
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>