Bug 477362 - Remove DIActivator
By removing the DIActivator and replacing it with a simple look-up in the
case of the DebugOptions, and moving the ServiceTracker to the LogHelper
class, it's possible to remove the DIActivator. Note that the ServiceTracker
does not need to be closed at the end of the bundle as this will be implicit.

Change-Id: I49004a40b763dadbdac55f2247f60221f17909fa
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>