Bug 466500: [GTK3] Trying to resize quick outline or quick type hierarchy closes the popup
In PopupDialog, Shells with no children are closed on deactivation. On
Linux-Gtk, this causes popups like the "Quick Outline" or "Quick
Type Hierarchy" to close prematurely when opening the system menu,
re-sizing, or moving the window with the mouse.

It is for this reason that we need to exclude the call to asyncClose()
from running on Gtk. This also fixes bug 113577.

Tested on Gtk3.16.6, and Gtk2.24.

Change-Id: I4111ad961e2b4e57c90939e7cb0e982b7bf3b894
Signed-off-by: Eric Williams <ericwill@redhat.com>