Bug 425525 - [ViewMgmt] Properties View does not populate with workbench selection when first opened and activated
Since on properties view opening the active view is the property view
itself, we can't use it to populate initial selection. 

Instead try to check active editor (if any) and use it's selection (if
not empty) for populating the view.

If there is no active editor, or it has no suitable selection, check
opened and visible views if there is exact one containing some
interesting selection. If there are multiple, don't use any of them (how
we should know which one is more important)?

Change-Id: I020c876194b8ec08b9aa847df0912a685da60bbb
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>