Bug 425525 - [ViewMgmt] Properties View does not populate with workbench selection when first opened and activated
While properties view is opening, the page active view is the property
view itself, so we can't use it to compute the originating "interesting"
part and populate initial selection.

Fortunately, the current page selection is still provided by the last
active part (view or editor). So we compare the active editor and all
visible views if the selection matches - and the one which has exact the
expected selection is the last active part. If this part is interesting,
it will be used to initially populate the selection.

Change-Id: Ic6899f129f6dc81c1086c2fb9aed58f1913bf739
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>